# docs/swagger.yaml
definitions:
  auth.LoginRequest:
    properties:
      password:
        example: password123
        type: string
      username:
        example: user123
        type: string
    required:
    - password
    - username
    type: object
  auth.RegisterRequest:
    properties:
      email:
        example: user@example.com
        type: string
      password:
        example: password123
        maxLength: 32
        minLength: 8
        type: string
      username:
        example: user123
        maxLength: 16
        minLength: 3
        type: string
    required:
    - email
    - password
    - username
    type: object
  wallet.ChangeBalanceRequest:
    properties:
      amount:
        type: number
      currency:
        enum:
        - USD
        - EUR
        - RUB
        type: string
    required:
    - currency
    type: object
  wallet.CurrenciesResponse:
    properties:
      EUR:
        type: number
      RUB:
        type: number
      USD:
        type: number
    type: object
  wallet.ExchangeRequest:
    properties:
      amount:
        type: number
      from_currency:
        enum:
        - USD
        - EUR
        - RUB
        type: string
      to_currency:
        enum:
        - USD
        - EUR
        - RUB
        type: string
    required:
    - from_currency
    - to_currency
    type: object
info:
  contact: {}
  title: Wallet service API
  version: 1.0.0
paths:
  /api/v1/auth/login/:
    post:
      consumes:
      - application/json
      description: Authenticate a user and return a JWT token
      parameters:
      - description: Login request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: JWT token
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid request or validation failed
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Invalid username or password
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: User login
      tags:
      - auth
  /api/v1/auth/register/:
    post:
      consumes:
      - application/json
      description: Register a new user with email, username, and password
      parameters:
      - description: Registration request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.RegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: User registered successfully
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Register a new user
      tags:
      - auth
  /api/v1/exchange/:
    post:
      consumes:
      - application/json
      description: Exchange a specified amount from one currency to another
      parameters:
      - default: Bearer <token>
        description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Exchange request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/wallet.ExchangeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Validation failed or invalid request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: user not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Exchange currency
      tags:
      - exchange
  /api/v1/exchange/rates/:
    get:
      consumes:
      - application/json
      description: Retrieve the latest exchange rates for supported currencies
      parameters:
      - default: Bearer <token>
        description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get exchange rates
      tags:
      - exchange
  /api/v1/wallet/balance/:
    get:
      consumes:
      - application/json
      description: Retrieve the balance of the user's wallet
      parameters:
      - default: Bearer <token>
        description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/wallet.CurrenciesResponse'
      summary: Get wallet balance
      tags:
      - wallet
  /api/v1/wallet/deposit/:
    post:
      consumes:
      - application/json
      description: Add a specified amount to the user's wallet
      parameters:
      - default: Bearer <token>
        description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Deposit request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/wallet.ChangeBalanceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Validation failed
          schema:
            additionalProperties: true
            type: object
        "401":
          description: userID not found in context
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Deposit money into wallet
      tags:
      - wallet
  /api/v1/wallet/withdraw/:
    post:
      consumes:
      - application/json
      description: Deduct a specified amount from the user's wallet
      parameters:
      - default: Bearer <token>
        description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Withdraw request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/wallet.ChangeBalanceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Validation failed
          schema:
            additionalProperties: true
            type: object
        "401":
          description: userID not found in context
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Withdraw money from wallet
      tags:
      - wallet
swagger: "2.0"